Make outl-md safe to share over iCloud
reconcile_md only emitted Create/Move/SetProp ops, never Op::Edit. Peers replayed empty blocks, regenerated .md from that empty state, and iCloud synced the empty file back. Every text edit turned into a silent delete across devices.

Walk the new AST in lockstep with the sidecar block list and emit one Op::Edit per block whose text doesn't match the workspace. Drop the leading dot from sidecar filenames (iOS Files.app hides dotted entries) and add resolve_sidecar_path to migrate older .foo.outl files on first read. Extract pure AST helpers (count_todos, path navigation) into outl_md::outline_ops so the mobile client can share them with the TUI.
