+++ crates/outl-md/src/reconcile.rs
@@ -10,11 +10,11 @@
 
 use crate::diff::diff_to_ops;
 use crate::matching::match_blocks;
-use crate::parse::parse;
+use crate::parse::{parse, OutlineNode};
 use crate::sidecar::{self, file_hash, sidecar_path_for, Sidecar, SidecarBlock, SIDECAR_VERSION};
 use outl_core::hlc::HlcGenerator;
 use outl_core::id::NodeId;
-use outl_core::op::LogOp;
+use outl_core::op::{LogOp, Op};
 use outl_core::workspace::{Workspace, WorkspaceError};
 use std::fs;
 use std::io::{self, Write};
@@ -129,6 +129,25 @@ pub fn reconcile_md(
         ops_applied += 1;
     }
 
+    // Synchronise block text with the workspace.
+    //
+    // `diff_to_ops` only knows about tree structure (Create / Move /
+    // SetProp). It never emits `Op::Edit` because computing the Yrs
+    // delta needs the live workspace, which isn't in its scope. The
+    // result is a tree of nodes that exist but have empty text — fine
+    // when the only consumer is the local sidecar (which carries the
+    // content hash) but **catastrophic across devices**: a peer
+    // replaying the op log materialises empty blocks, regenerates
+    // `.md` from that empty state, and iCloud syncs the empty `.md`
+    // back to us. Every text edit silently turns into a deletion.
+    //
+    // Fix: walk the new AST in lockstep with the freshly built
+    // sidecar block list (same DFS preorder) and emit one
+    // `Op::Edit` per block whose text doesn't match what the
+    // workspace already has. Idempotent: `build_text_replace_update`
+    // returns an empty update when text is unchanged.
+    ops_applied += sync_block_text(ws, hlc, &new_ast.blocks, &plan.new_sidecar.blocks)?;
+
     let new_sidecar = Sidecar {
         version: SIDECAR_VERSION,
         page_id,
@@ -144,6 +163,58 @@ pub fn reconcile_md(
         orphans: orphans.len(),
         created_sidecar,
     })
+}
+
+/// Walk the parsed AST and the freshly built sidecar block list in
+/// lockstep (both in DFS preorder) and emit one `Op::Edit` per block
+/// whose text doesn't already match what's in the workspace.
+///
+/// Returns the number of `Op::Edit` ops applied. Idempotent: skips
+/// blocks whose text already matches (the Yrs delta would be empty).
+fn sync_block_text(
+    ws: &mut Workspace,
+    hlc: &HlcGenerator,
+    ast_blocks: &[OutlineNode],
+    sidecar_blocks: &[SidecarBlock],
+) -> Result<usize, WorkspaceError> {
+    let mut idx = 0usize;
+    let mut applied = 0usize;
+    walk_text_sync(ws, hlc, ast_blocks, sidecar_blocks, &mut idx, &mut applied)?;
+    Ok(applied)
+}
+
+fn walk_text_sync(
+    ws: &mut Workspace,
+    hlc: &HlcGenerator,
+    ast_blocks: &[OutlineNode],
+    sidecar_blocks: &[SidecarBlock],
+    idx: &mut usize,
+    applied: &mut usize,
+) -> Result<(), WorkspaceError> {
+    for block in ast_blocks {
+        if let Some(entry) = sidecar_blocks.get(*idx) {
+            let node = entry.id;
+            let current = ws.block_text(node).unwrap_or_default();
+            if current != block.text {
+                let update = ws.build_text_replace_update(node, &block.text);
+                if !update.is_empty() {
+                    let ts = hlc.next();
+                    ws.apply(LogOp {
+                        ts,
+                        actor: ts.actor,
+                        op: Op::Edit {
+                            node,
+                            text_op: update,
+                        },
+                    })?;
+                    *applied += 1;
+                }
+            }
+        }
+        *idx += 1;
+        walk_text_sync(ws, hlc, &block.children, sidecar_blocks, idx, applied)?;
+    }
+    Ok(())
 }
 
 fn log_orphans(
